change type of min,max,suggestedMin and suggestedMax to Double

+++ src/main/java/com/byteowls/vaadin/chartjs/options/scale/LinearTicks.java
@@ -12,13 +12,13 @@ public class LinearTicks<T> extends Ticks<T> implements JsonBuilder {
     private static final long serialVersionUID = -7751881366365153424L;
 
     private Boolean beginAtZero;
-    private Integer min;
-    private Integer max;
+    private Double min;
+    private Double max;
     private Integer maxTicksLimit;
     private Double fixedStepSize;
     private Double stepSize;
-    private Integer suggestedMax;
-    private Integer suggestedMin;
+    private Double suggestedMax;
+    private Double suggestedMin;
 
     public LinearTicks(T parent) {
         super(parent);
@@ -35,7 +35,7 @@ public class LinearTicks<T> extends Ticks<T> implements JsonBuilder {
     /**
      * User defined minimum number for the scale, overrides minimum value from data.
      */
-    public LinearTicks<T> min(int min) {
+    public LinearTicks<T> min(Double min) {
         this.min = min;
         return this;
     }
@@ -43,7 +43,7 @@ public class LinearTicks<T> extends Ticks<T> implements JsonBuilder {
     /**
      * User defined maximum number for the scale, overrides maximum value from data.
      */
-    public LinearTicks<T> max(int max) {
+    public LinearTicks<T> max(Double max) {
         this.max = max;
         return this;
     }
@@ -76,7 +76,7 @@ public class LinearTicks<T> extends Ticks<T> implements JsonBuilder {
     /**
      * User defined maximum number for the scale, overrides maximum value except for if it is lower than the maximum value.
      */
-    public LinearTicks<T> suggestedMax(int suggestedMax) {
+    public LinearTicks<T> suggestedMax(Double suggestedMax) {
         this.suggestedMax = suggestedMax;
         return this;
     }
@@ -84,7 +84,7 @@ public class LinearTicks<T> extends Ticks<T> implements JsonBuilder {
     /**
      * User defined minimum number for the scale, overrides minimum value except for if it is higher than the minimum value.
      */
-    public LinearTicks<T> suggestedMin(int suggestedMin) {
+    public LinearTicks<T> suggestedMin(Double suggestedMin) {
         this.suggestedMin = suggestedMin;
         return this;
     }
